instrument specific registers in their separate module
