+++ CMakeLists.txt
@@ -29,6 +29,7 @@ add_executable(pico_scpi_usbtmc_labtool
         ${CMAKE_CURRENT_SOURCE_DIR}/source/usb/usbtmc_app.c
         ${CMAKE_CURRENT_SOURCE_DIR}/source/usb/usbtmc_device_custom.c
         ${CMAKE_CURRENT_SOURCE_DIR}/source/scpi/scpi-def.c
+        ${CMAKE_CURRENT_SOURCE_DIR}/source/scpi/registers/registers.c
         ${CMAKE_CURRENT_SOURCE_DIR}/source/gpio/gpio_utils.c
         ${CMAKE_CURRENT_SOURCE_DIR}/source/adc/adc_utils.c
         ${CMAKE_CURRENT_SOURCE_DIR}/source/i2c/i2c_utils.c
